requirements.txt
aiohttp>=3.12.15
betterproto>=1.2.5
mini-racer>=0.12.4
yarl>=1.20.1
